[refactor] Retrofit과 OkHttpClient 분리
