+++ Android/FunBox/app/src/main/java/com/rpg/funbox/data/RetrofitInstance.kt
@@ -2,8 +2,6 @@ package com.rpg.funbox.data
 
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 
@@ -14,19 +12,10 @@ object RetrofitInstance {
         .add(KotlinJsonAdapterFactory())
         .build()
 
-    private val okHttpClient: OkHttpClient by lazy {
-        val interceptor = HttpLoggingInterceptor()
-            .setLevel(HttpLoggingInterceptor.Level.BODY)
-        OkHttpClient.Builder().run {
-            this.addInterceptor(interceptor)
-            this.build()
-        }
-    }
-
     val retrofit: Retrofit by lazy {
         Retrofit.Builder()
             .baseUrl(FUNBOX_BASE_URL)
-            .client(okHttpClient)
+            .client(OkHttpClientInstance.okHttpClient)
             .addConverterFactory(MoshiConverterFactory.create(moshi))
             .build()
     }
